Move to pytests, add test for CryptSvce_manual
